refactor(skills): inline the skills page into its route
`SkillsPage` was a single-consumer, eight-prop pass-through with no test file of its own, so the extraction bought nothing that the route did not already have.
Every sibling data-owning settings route (`agent-settings`, `llm-settings`, `app-settings`, `mcp`) keeps its hooks, state, and JSX inline, and this route lands well under their size.
